+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased version -- v0.13.1
 ## Added
 - New script to parse download requests from S3 log files for UMP books 
+- New script to update current UofM manifests with fulfill endpoints to replace pdf/epub urls
 ## Fixed
 
 ## 2024-03-21 -- v0.13.0